+++ projects/myrmidon/cadmus-profile-import/src/public-api.ts
@@ -1,0 +1,9 @@
+/*
+ * Public API Surface of cadmus-profile-import
+ */
+
+export * from './lib/facet-import/facet-import.component';
+export * from './lib/facet-import-page/facet-import-page.component';
+export * from './lib/settings-import/settings-import.component';
+export * from './lib/settings-import-page/settings-import-page.component';
+export * from './lib/cadmus-profile-import.routes';